BREAKING: use govuk for forms
We've decided to move all the forms from Koi to content so we don't
have to test and maintain two versions. We've never found a situation
where we want to use content without govuk so this seems like a
reasonable compromise.

See Koi or dummy for a standard Katalyst CSS setup.